local_kafka/requirements.txt
kafka-python==2.0.2
pytz==2021.1
